Move discovered files collection.
