--- BFBMX.Service/Collections/DiscoveredFilesCollection.cs
@@ -1,38 +1,0 @@
-﻿using BFBMX.Service.Models;
-using System.Collections.Concurrent;
-
-namespace BFBMX.Service.Collections
-{
-    public interface IDiscoveredFilesCollection
-    {
-        int MaxItems { get; }
-
-        void Enqueue(DiscoveredFileModel item);
-    }
-
-    /// <summary>
-    /// Concurrent Queue datastructure with a maximum item limit.
-    /// </summary>
-    public class DiscoveredFilesCollection : ConcurrentQueue<DiscoveredFileModel>, IDiscoveredFilesCollection
-    {
-        public int MaxItems { get; } = 6;
-
-        public DiscoveredFilesCollection()
-        {
-        }
-
-        public new void Enqueue(DiscoveredFileModel item)
-        {
-            base.Enqueue(item);
-
-            while (base.Count > this.MaxItems)
-            {
-                base.TryDequeue(out DiscoveredFileModel? _);
-            }
-        }
-
-        public DiscoveredFilesCollection(IEnumerable<DiscoveredFileModel> collection) : base(collection)
-        {
-        }
-    }
-}
